Moving the memory leak test for SynchronousQueue into its own test and runs only for JDK20+, using VirtualThreads
